testing

+++ src/main/java/com/square/getmedoc/db/model/Address.java
@@ -1,6 +1,6 @@
 package com.square.getmedoc.db.model;
 
-// Generated Aug 29, 2013 7:53:12 PM by Hibernate Tools 4.0.0
+// Generated Sep 2, 2013 7:42:39 PM by Hibernate Tools 4.0.0
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -19,7 +19,7 @@ import javax.persistence.Table;
 @Table(name = "address", catalog = "getmedoc")
 public class Address implements java.io.Serializable {
 
-	private Integer addressid;
+	private Long addressid;
 	private Appuser appuser;
 	private String line1;
 	private String line2;
@@ -59,11 +59,11 @@ public class Address implements java.io.Serializable {
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
 	@Column(name = "addressid", unique = true, nullable = false)
-	public Integer getAddressid() {
+	public Long getAddressid() {
 		return this.addressid;
 	}
 
-	public void setAddressid(Integer addressid) {
+	public void setAddressid(Long addressid) {
 		this.addressid = addressid;
 	}
